Address first round of review suggestions.
